latest generated openwire.net code which now compiles OK!
git-svn-id: https://svn.apache.org/repos/asf/incubator/activemq/trunk@367311 13f79535-47bb-0310-9956-ffa450edef68

+++ openwire-dotnet/src/OpenWire.Core/AbstractCommand.cs
@@ -23,7 +23,7 @@ namespace OpenWire.Core
 			
 		}
 		
-         public virtual int GetCommandType()
+         public virtual byte GetCommandType()
          {
          	    return 0;
  	    }
